Add files via upload
Alot of work and tweaking still to be done. Still have to link the PentestGPT output to the prompt generation. Might also need fine tuning of a GGUF model such as spicyboros or an uncensored vicuna...... Credit to GPT4ALL and PentestGPT for inspiration, just have to stitch multiple different projects together.